Use DI for message commands
