+++ FrEee.Core.Domain/Gameplay/Commands/Messages/IDeleteMessageCommand.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace FrEee.Gameplay.Commands.Messages;
+public interface IDeleteMessageCommand
+	: IMessageCommand
+{
+}
